Combine static methods into RawRepresentable FilePermissions
This seems like more idiomatic Swift and keeps the code that converts from the raw number value local to the rest of FilePermissions code.

The decision to make the RawValue type Int32 was motivated by the LibSSH2 permissions constants having that type instead of mode_t/UInt16. Since Shout is based on LibSSH2 and the use of FilePermissions will likely be used with it's permission values internally, this might avoid tedious conversion between integer types. If the opposite ends up being the case, maybe the RawValue type should be changed to mode_t/UInt16 instead.