Consolidate Bindings target into Shout target
The motivation for this change is because the division in responsibilities and dependencies was not clear and introduced difficulty when trying to extract the FilePermissions type for general use in both targets.

The Bindings source was moved into the Shout target directory and made internal, except for SFTP and any types that needed to be public because they were used in Shout's existing public types.